+++ cli.js
@@ -6,14 +6,22 @@ const process = require('process');
 const packageJson = require('./package.json');
 const bml = require('./bml.js');
 
+const SEED_RE = /^-?\d+$/;
 
 const HELP_SWITCHES = ['-h', '--h', '-help', '--help'];
 const VERSION_SWITCHES = ['-v', '--version'];
-const SEED_SWITCHES = ['--seed']
+const SEED_SWITCHES = ['--seed'];
+const NO_EVAL_SWITCHES = ['--no-eval'];
+const RENDER_MARKDOWN_SWITCHES = ['--render-markdown'];
+const NO_WHITESPACE_CLEANUP_SWITCHES = ['--no-whitespace-cleanup'];
+
+const ALL_SWITCHES = [].concat(
+  HELP_SWITCHES, VERSION_SWITCHES,
+  SEED_SWITCHES, NO_EVAL_SWITCHES,
+  RENDER_MARKDOWN_SWITCHES, NO_WHITESPACE_CLEANUP_SWITCHES);
 
 
-function readFromStdin(seed) {
-  let settings = seed ? { randomSeed: seed } : {};
+function readFromStdin(settings) {
   return {
     bmlSource: fs.readFileSync(0, 'utf8'), // STDIN_FILENO = 0
     settings
@@ -21,12 +29,11 @@ function readFromStdin(seed) {
 }
 
 
-function readFromPath(path, seed) {
+function readFromPath(path, settings) {
   if (!fs.existsSync(path)) {
     handleNonexistingPath(path);
     process.exit(1);
   }
-  let settings = seed ? { randomSeed: seed } : {};
   return {
     bmlSource: '' + fs.readFileSync(path),
     settings
@@ -52,12 +59,15 @@ function printHelp() {
 
     Options which, if present, should be the only options given:
 
-    ${HELP_SWITCHES}    print this help and quit
-    ${VERSION_SWITCHES}           print the bml version number and quiet
+    ${HELP_SWITCHES}        print this help and quit
+    ${VERSION_SWITCHES}               print the BML version number and quiet
 
     Other options:
 
-    ${SEED_SWITCHES} INTEGER         set the random seed for the bml render
+    ${SEED_SWITCHES} INTEGER             set the random seed for the bml render
+    ${NO_EVAL_SWITCHES}                  disable Javascript evaluation
+    ${RENDER_MARKDOWN_SWITCHES}          render the document as markdown to HTML
+    ${NO_WHITESPACE_CLEANUP_SWITCHES}    disable whitespace cleanup
 
   Source Code at https://github.com/ajyoon/bml
   Report Bugs at https://github.com/ajyoon/bml/issues
@@ -66,11 +76,26 @@ function printHelp() {
   );
 }
 
+// The way this function is just a passthrough really illustrates
+// why the higher-level-function approach of this module is awkward
+function printHelpForError() {
+  printHelp();
+}
+
 
 function printVersionInfo() {
   process.stdout.write(packageJson.version + '\n');
 }
 
+function argsContainAnyUnknownSwitches(args) {
+  let unknown_arg = args.find(
+    (arg) => !SEED_RE.test(arg) && arg.startsWith('-') && !ALL_SWITCHES.includes(arg));
+  if (unknown_arg) {
+    console.error(`Unknown argument ${unknown_arg}`);
+    return true;
+  }
+  return false;
+}
 
 /**
  * Parse the given command line arguments and determine the action needed.
@@ -80,62 +105,81 @@ function printVersionInfo() {
  * @return {Object} of the form {function: Function, args: ...Any}
  */
 function determineAction(args) {
-  let defaultAction = {
-    function: printHelp,
-    args: [],
+  let errorAction = {
+    function: printHelpForError,
+    args: []
   };
+  
+  if (argsContainAnyUnknownSwitches(args)) {
+    return errorAction;
+  }
 
-  switch (args.length) {
-  case 0:
-    return {
-      function: readFromStdin,
-      args: [],
-    };
-
-  case 1:
-    if (HELP_SWITCHES.indexOf(args[0]) !== -1) {
-      return defaultAction;
-    } else if (VERSION_SWITCHES.indexOf(args[0]) !== -1) {
+  let expectSeed = false;
+  
+  let file = null;
+  let noEval = false;
+  let renderMarkdown = false;
+  let noWhitespaceCleanup = false;
+  let seed = null;
+  
+  for (let arg of args) {
+    if (expectSeed) {
+      if (SEED_RE.test(arg)) {
+        seed = Number(arg);
+        expectSeed = false;
+      } else {
+        console.error('Invalid seed: ' + arg);
+        return errorAction;
+      }
+    } else if (HELP_SWITCHES.includes(arg)) {
+      return {
+        function: printHelp,
+        args: []
+      };
+    } else if (VERSION_SWITCHES.includes(arg)) {
       return {
         function: printVersionInfo,
         args: [],
       };
+    } else if (NO_EVAL_SWITCHES.includes(arg)) {
+      noEval = true;
+    } else if (RENDER_MARKDOWN_SWITCHES.includes(arg)) {
+      renderMarkdown = true;
+    } else if (NO_WHITESPACE_CLEANUP_SWITCHES.includes(arg)) {
+      noWhitespaceCleanup = true;
+    } else if (SEED_SWITCHES.includes(arg)) {
+      expectSeed = true;
     } else {
-      return {
-        function: readFromPath,
-        args: [args[0]],
-      };
+      if (file !== null) {
+        console.error('More than one path provided.');
+        return errorAction;
+      }
+      file = arg;
     }
-  case 2:
-    // Only really handle the case where a seed is given and bml
-    // is taken from stdin
-    if (SEED_SWITCHES.indexOf(args[0]) !== -1) {
-      let seedNum = Number(args[1]);
-      if (isNaN(seedNum) || !Number.isInteger(seedNum)) {
-        return defaultAction;
-      }
-      return {
-        function: readFromStdin,
-        args: [seedNum]
-      }
+  }
+  
+  if (expectSeed) {
+    console.error('No seed provided.')
+    return errorAction;
+  }
+
+  let settings = {
+    randomSeed: seed,
+    allowEval: !noEval,
+    renderMarkdown: renderMarkdown,
+    whitespaceCleanup: !noWhitespaceCleanup
+  };
+  
+  if (file === null) {
+    return {
+      function: readFromStdin,
+      args: [settings]
     }
-    return defaultAction;
-  case 3:
-    // Only really handle the case where a seed and path are given
-    // e.g. `bml --seed 1234 ./someFile.bml`
-    if (SEED_SWITCHES.indexOf(args[0]) !== -1) {
-      let seedNum = Number(args[1]);
-      if (isNaN(seedNum) || !Number.isInteger(seedNum)) {
-        return defaultAction;
-      }
-      return {
-        function: readFromPath,
-        args: [args[2], seedNum]
-      }
+  } else {
+    return {
+      function: readFromPath,
+      args: [file, settings]
     }
-    return defaultAction;
-  default:
-    return defaultAction;
   }
 }
 
@@ -163,12 +207,12 @@ function main() {
   let strippedArgs = stripArgs(process.argv);
   let action = determineAction(strippedArgs);
 
-  if (action.function === printHelp) {
-    printHelp();
+  if (action.function === printHelp || action.function === printVersionInfo) {
+    action.function();
     process.exit(0);
-  } else if (action.function === printVersionInfo) {
-    printVersionInfo();
-    process.exit(0);
+  } else if (action.function == printHelpForError) {
+    action.function();
+    process.exit(1);
   } else {
     let { bmlSource, settings } = action.function(...action.args);
     let renderedContent = runBmlWithErrorCheck(bmlSource, settings);
@@ -188,6 +232,7 @@ exports.VERSION_SWITCHES = VERSION_SWITCHES;
 exports.readFromStdin = readFromStdin;
 exports.readFromPath = readFromPath;
 exports.printHelp = printHelp;
+exports.printHelpForError = printHelpForError;
 exports.printVersionInfo = printVersionInfo;
 exports.determineAction = determineAction;
 exports.stripArgs = stripArgs;