+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ### 0.0.23-dev:
-* tbd...
+* Make all render settings available to the CLI
+* Make CLI errors from invalid arguments log more useful messages to
+  stderr and give exit code 1.
 
 ### 0.0.22:
 * Remove accidentally-left-in debug log on regex matcher parsing